@@ -1,24 +1,6 @@
 {
-  "name": "angular",
+  "name": "test-ap",
   "version": "0.0.0",
-  "private": true,
-  "dependencies": {
-    "@angular/animations": "^10.0.0",
-    "@angular/cdk": "7.2.1",
-    "@angular/common": "^10.0.0",
-    "@angular/compiler": "^10.0.0",
-    "@angular/core": "^10.0.0",
-    "@angular/forms": "^10.0.0",
-    "@angular/material": "7.2.1",
-    "@angular/platform-browser": "^10.0.0",
-    "@angular/platform-browser-dynamic": "^10.0.0",
-    "@angular/router": "^10.0.0",
-    "ng2-charts": "1.6.0",
-    "core-js": "^3.6.5",
-    "rxjs": "^6.5.5",
-    "tslib": "^2.0.0",
-    "zone.js": "^0.10.3"
-  },
   "scripts": {
     "ng": "ng",
     "start": "ng serve",
@@ -27,25 +9,42 @@
     "lint": "ng lint",
     "e2e": "ng e2e"
   },
+  "private": true,
+  "dependencies": {
+    "@angular/animations": "~10.1.0-next.4",
+    "@angular/cdk": "^10.1.2",
+    "@angular/common": "~10.1.0-next.4",
+    "@angular/compiler": "~10.1.0-next.4",
+    "@angular/core": "~10.1.0-next.4",
+    "@angular/forms": "~10.1.0-next.4",
+    "@angular/material": "^7.2.1",
+    "@angular/platform-browser": "~10.1.0-next.4",
+    "@angular/platform-browser-dynamic": "~10.1.0-next.4",
+    "@angular/router": "~10.1.0-next.4",
+    "chart.js": "^2.9.3",
+    "ng2-charts": "^2.3.2",
+    "rxjs": "~6.6.0",
+    "tslib": "^2.0.0",
+    "zone.js": "~0.10.2"
+  },
   "devDependencies": {
-    "@angular-devkit/build-angular": "~0.10.0",
-    "@angular/cli": "~7.0.2",
-    "@angular/compiler-cli": "~7.0.0",
-    "@angular/language-service": "~7.0.0",
-    "@types/node": "~8.9.4",
-    "@types/jasmine": "~2.8.8",
+    "@angular-devkit/build-angular": "~0.1001.0-next.4",
+    "@angular/cli": "~10.1.0-next.4",
+    "@angular/compiler-cli": "~10.1.0-next.4",
+    "@types/node": "^12.11.1",
+    "@types/jasmine": "~3.5.0",
     "@types/jasminewd2": "~2.0.3",
-    "codelyzer": "~4.5.0",
-    "jasmine-core": "~2.99.1",
-    "jasmine-spec-reporter": "~4.2.1",
-    "karma": "~3.0.0",
-    "karma-chrome-launcher": "~2.2.0",
-    "karma-coverage-istanbul-reporter": "~2.0.1",
-    "karma-jasmine": "~1.1.2",
-    "karma-jasmine-html-reporter": "^0.2.2",
-    "protractor": "~5.4.0",
-    "ts-node": "~7.0.0",
-    "tslint": "~5.11.0",
-    "typescript": "~3.1.1"
+    "codelyzer": "^6.0.0",
+    "jasmine-core": "~3.6.0",
+    "jasmine-spec-reporter": "~5.0.0",
+    "karma": "~5.0.0",
+    "karma-chrome-launcher": "~3.1.0",
+    "karma-coverage-istanbul-reporter": "~3.0.2",
+    "karma-jasmine": "~3.3.0",
+    "karma-jasmine-html-reporter": "^1.5.0",
+    "protractor": "~7.0.0",
+    "ts-node": "~8.3.0",
+    "tslint": "~6.1.0",
+    "typescript": "~3.9.5"
   }
 }
